NF: separate test setup to mock mod
