+++ frame/timestamp/src/lib.rs
@@ -103,6 +103,8 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 #[cfg(test)]
+mod mock;
+#[cfg(test)]
 mod tests;
 mod benchmarking;
 pub mod weights;
